+++ Gateway/Response/ResponseHandler.php
@@ -52,6 +52,7 @@ class ResponseHandler implements HandlerInterface
 
         $payment->setTransactionId($almaPayment->id);
         $payment->setAdditionalInformation(Config::ORDER_PAYMENT_URL, $almaPayment->url);
+        $payment->setAdditionalInformation(Config::ORDER_PAYMENT_TRIGGER, $almaPayment->deferred_trigger);
         $payment->setIsTransactionClosed(false);
 
         $order = $payment->getOrder();