+++ view/frontend/web/js/view/payment/method-renderer/alma_monthly_payments.js
@@ -161,6 +161,16 @@ define(
                 }
                 return label;
             },
+            triggerIsEnable: function() {
+                return window.checkoutConfig.payment[this.item.method].triggerEnable;
+            },
+            onTriggerLabel: function(numInstallment) {
+                var label = $t('%1 month later').replace('%1', numInstallment);
+                if (numInstallment==0){
+                    label = window.checkoutConfig.payment[this.item.method].triggerLabel;
+                }
+                return label;
+            },
 
             formattedDate: function (ts) {
                 return (new Date(ts * 1000)).toLocaleDateString(this.config.locale);
@@ -171,7 +181,8 @@ define(
             },
 
             cartTotal: function () {
-                return priceUtils.formatPrice(this.totals().grand_total, window.checkoutConfig.priceFormat);
+                console.log(this.totals())
+                return priceUtils.formatPrice(this.totals().base_grand_total, window.checkoutConfig.priceFormat);
             },
 
             customerTotalCostAmount: function (cost) {